backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
PyYAML
